Remove all usage of legacy
