+++ packages/cf.js/src/index.ts
@@ -1,5 +1,4 @@
 import { AppFactory } from "./app-factory";
-import * as legacy from "./legacy";
 import { Provider } from "./provider";
 import * as types from "./types";
 import * as utils from "./utils";
@@ -7,11 +6,10 @@ import * as utils from "./utils";
 const cf = {
   AppFactory,
   Provider,
-  legacy,
   types,
   utils
 };
 
-export { AppFactory, Provider, legacy, types, utils };
+export { AppFactory, Provider, types, utils };
 
 export default cf;
