--- packages/cf.js/src/legacy/index.ts
@@ -1,7 +1,0 @@
-import * as app from "./app";
-import * as channel from "./channel";
-import { NotificationType } from "./mixins/observable";
-import * as node from "./node";
-import * as utils from "./utils";
-
-export { app, channel, node, NotificationType, utils };
