--- packages/cf.js/src/legacy/mixins/apply.ts
@@ -1,9 +1,0 @@
-// https://www.typescriptlang.org/docs/handbook/mixins.html
-
-export function applyMixins(derivedCtor: any, baseCtors: any[]) {
-  baseCtors.forEach(baseCtor => {
-    Object.getOwnPropertyNames(baseCtor.prototype).forEach(name => {
-      derivedCtor.prototype[name] = baseCtor.prototype[name];
-    });
-  });
-}
