+++ packages/cf.js/src/utils/index.ts
@@ -1,5 +1,6 @@
 import * as abi from "./abi";
+import { generateAbiEncodings } from "./encodings";
 
-export { abi };
+export { abi, generateAbiEncodings };
 
 export * from "./signature";
